requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
dateparser==1.1.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pycparser==2.21
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
six==1.16.0
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
